requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.17.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.1.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Django==5.1.2
django-jazzmin==3.0.1
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
Flask==3.0.3
Flask-Cors==5.0.0
frozenlist==1.5.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iso8601==1.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.9
MarkupSafe==2.1.5
mpremote==1.24.1
multidict==6.1.0
numpy==2.2.2
packaging==24.1
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.4
pydantic-settings==2.3.3
pydantic_core==2.18.4
pypika-tortoise==0.1.6
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-jose==3.3.0
python-multipart==0.0.20
python-socketio==5.12.1
python-telegram-bot==21.4
pytz==2024.1
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlparse==0.5.1
starlette==0.45.3
tortoise-orm==0.21.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
websockets==14.2
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.18.3
